Consolidate memory servers
